+++ src/main/java/org/example/eticaret/service/ProductService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProductService {
 
     ProductDto save(ProductDto productDto);
-    ProductDto getProduct(Long id);
+    ProductDto getProduct(int id);
     List<ProductDto> getAll();
     ProductDto update(Long id, ProductDto productDto);
     void  delete(Long id);
